Changes position of node logging test
